% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{association}
\alias{association}
\alias{association,WordAssociation-method}
\title{Get the association strengh for all association computed in this \code{\link{WordAssociation-class}} object}
\usage{
association(obj)

\S4method{association}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
a numeric vector
}
\description{
Get the association strengh for all association computed in this \code{\link{WordAssociation-class}} object
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood", "specificities"))
indicator.name(x)
}
