% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{bptest}
\alias{bptest}
\title{bptest}
\usage{
bptest(formula, varformula = NULL, studentize = TRUE, data = list())
}
\arguments{
\item{formula}{a symbolic description for the model to be tested (or a fitted "lm" object)}

\item{varformula}{a formula describing only the potential explanatory variables for the variance (no dependent variable needed). By default the same explanatory variables are taken as in the main regression model.}

\item{studentize}{logical. If set to TRUE Koenker's studentized version of the test statistic will be used.}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which bptest is called from.}
}
\value{
A list with class "htest" containing a statistic (the value of the test statistic), p.value (the p-value of the test), parameter (degrees of freedom), method (a character string indicating what type of test was performed), and data.name(a character string giving the name(s) of the data)
}
\description{
This function allows users to apply the Breusch-Pagan test for heteroskedasticity on their regression. The Breusch-Pagan test fits a linear regression model to the residuals of a linear regression model (by default the same explanatory variables are taken as in the main regression model) and rejects if too much of the variance is explained by the additional explanatory variables.
}
\examples{
generate a regressor
x <- rep(c(-1,1), 50)
generate heteroskedastic and homoskedastic disturbances
err1 <- rnorm(100, sd=rep(c(1,2), 50))
err2 <- rnorm(100)
generate a linear relationship
y1 <- 1 + x + err1
y2 <- 1 + x + err2
perform Breusch-Pagan test
bptest(y1 ~ x)
bptest(y2 ~ x)

}
