% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{coeftest}
\alias{coeftest}
\title{coeftest}
\usage{
coeftest(x, vcov. = NULL, df = NULL, ...)
}
\arguments{
\item{x}{an object}

\item{df}{degrees of freedom: iff it's a finite positive number a t test with df degrees of freedom is performed, otherwise a z test is used}

\item{...}{further arguments passed to methods and vcov}

\item{vcov}{specification of covariance matrix of estimated coefficients}

\item{parm}{specification of which parameters are to be given confidence intervals, if missing all params are considered}

\item{level}{confidence level required}
}
\value{
An object with class "coeftest" which is a coefficient matrix with columns containing estimates, standard errors, test statistics, and p values.
}
\description{
This function allows users to apply t- and z-Wald tests on their coefficients
}
\examples{
generate a regressor
## a different covariance matrix can be also used:
(ct <- coeftest(fm, df = Inf, vcov = vcovHC))

}
