% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{export}
\alias{export}
\title{export}
\usage{
export(x, file)
}
\arguments{
\item{x}{the object itself}

\item{file}{the file to deposit the object into}
}
\value{
None
}
\description{
This function allows users to export objects into a desired file in their working directory
}
\examples{
export(df, 'test.csv')
}
