% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{heckit}
\alias{heckit}
\title{PseudoR2}
\usage{
heckit(selection, outcome, data = sys.frame(sys.parent()),
  method = "2step", ...)
}
\arguments{
\item{selection}{formula}

\item{outcome}{outcome equation, either single equation or list of two equations}

\item{data}{optional dataframe list or environment}

\item{method}{how to estimate the model - either ml for max likelihood or 2step for 2-step estimation, or model.frame for returning the model frame only}

\item{weights}{optional vector of prior weights}

\item{type}{model type}
}
\value{
value of desired statistic
}
\description{
This function allows users to execute Heckman selection models (tobits) with one or two outcomes
}
\examples{
usage
heckit( selection, outcome, data = sys.frame(sys.parent()),
method = "2step", ... )

}
