% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{import}
\alias{import}
\title{import}
\usage{
import(file)
}
\arguments{
\item{file}{the file to import}
}
\value{
None
}
\description{
This function allows importation of files in your directory into your R file
}
\examples{
import('test.csv')
}
