% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{indicator.name}
\alias{indicator.name}
\alias{indicator.name,WordAssociation-method}
\title{The names of the statistical indicators used in a \code{\link{WordAssociation-class}} object}
\usage{
indicator.name(obj)

\S4method{indicator.name}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
a character vector
}
\description{
The names of the statistical indicators used in a \code{\link{WordAssociation-class}} object
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood", "specificities"))
indicator.name(x)
}
