% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{ivreg}
\alias{ivreg}
\title{ivreg}
\usage{
ivreg(formula, instruments, data, subset, na.action, weights, offset,
  contrasts = NULL, model = TRUE, y = TRUE, x = FALSE, ...)
}
\arguments{
\item{formula}{formula specification of both formulas with instrument included in the second formula}

\item{instruments}{read above}

\item{data}{optional data frame containing variables in the model}

\item{subset}{optional vector specifying subset of observations to be used in fitting the model}

\item{na.action}{function indicating what should be done with NAs in data}

\item{weights}{optional vector of weights to be used in fitting process}

\item{offset}{optional offset}

\item{contrasts}{optional list}

\item{model}{logical, if true model frame is returned}

\item{y}{logical, if true response is returned}

\item{x}{logical, if true model matrices are returned}

\item{...}{further arguments}
}
\value{
ivreg returns an object of class "ivreg", with econometric model with intstrumental variable used
}
\description{
This function allows users to fit instrumental-variable regression through two-stage least squares. Regressors and instruments for ivreg are most easily specified in a formula with two parts on the right-hand side, e.g., y ~ x1 + x2 | z1 + z2 + z3, where x1 and x2 are the regressors and z1, z2, and z3 are the instruments. Note that exogenous regressors have to be included as instruments for themselves.
}
\examples{
modeling
fm <- ivreg(log(packs) ~ log(rprice) + log(rincome) | log(rincome) + tdiff + I(tax/cpi),
data = CigarettesSW, subset = year == "1995")
summary(fm)
summary(fm, vcov = sandwich, df = Inf, diagnostics = TRUE)

}
