% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{K-big}
\alias{K-big}
\alias{K}
\alias{K,WordAssociation-method}
\title{Total frequency of forms in a \code{\link{WordAssociation-class}} object}
\usage{
K(obj)

\S4method{K}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
a numeric vector
}
\description{
Total frequency of forms in a \code{\link{WordAssociation-class}} object
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
K(x)
}
