\name{k-methods}
\docType{methods}
\alias{k}
\alias{k-methods}
\alias{k,WordAssociation-method}
\title{get the subfrequency of a form in a \code{\link{WordAssociation}} object}
\description{

}
\section{Methods}{
\describe{

\item{\code{signature(obj = "WordAssociation")}}{

}
}}
\keyword{methods}

