% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{k-small}
\alias{k-small}
\alias{k}
\alias{k,WordAssociation-method}
\title{Subfrequencies of forms in a \code{\link{WordAssociation-class}} object.}
\usage{
k(obj)

\S4method{k}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
a numeric vector
}
\description{
Subfrequencies of forms in a \code{\link{WordAssociation-class}} object.
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
k(x)
}
