% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{margins}
\alias{margins}
\title{margins}
\usage{
margins(model, ...)
}
\arguments{
\item{model}{Desired model to marginify}

\item{...}{arguments passed to methods}
}
\value{
dataframe of class "margins" returned containing margins
}
\description{
This function allows users to access more descriptive model margins (like Stata's margins command)
}
\examples{
using a linear model
basic example using linear model
require("datasets")
x <- lm(mpg ~ cyl * hp + wt, data = head(mtcars))
margins(x)

}
