% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{N-big}
\alias{N-big}
\alias{N}
\alias{N,WordAssociation-method}
\title{Number of tokens in its corpus for each forms in a \code{\link{WordAssociation-class}} object.}
\usage{
N(obj)

\S4method{N}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
a numeric vector
}
\description{
Number of tokens in its corpus for each forms in a \code{\link{WordAssociation-class}} object.
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
N(x)
}
