% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{n-small}
\alias{n-small}
\alias{n}
\alias{n,WordAssociation-method}
\title{Number of tokens in its subcorpus for each forms in a \code{\link{WordAssociation-class}} object.}
\usage{
n(obj)

\S4method{n}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
a numeric vector
}
\description{
Number of tokens in its subcorpus for each forms in a \code{\link{WordAssociation-class}} object.
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
n(x)
}
