% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{parts}
\alias{parts}
\alias{parts,WordAssociation-method}
\title{The part of the corpus each entry belong to in a \code{\link{WordAssociation-class}} object}
\usage{
parts(obj)

\S4method{parts}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
a character vector
}
\description{
The part of the corpus each entry belong to in a \code{\link{WordAssociation-class}} object
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
parts(x)
}
