% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.print.R
\name{print,WordAssociation-method}
\alias{print,WordAssociation-method}
\title{Print a \code{\link{WordAssociation-class}} object}
\usage{
\S4method{print}{WordAssociation}(
  x,
  from = 1,
  to = 100,
  threshold = NULL,
  types = NULL,
  parts = NULL,
  sort.by = indicator.name(x)[1],
  file = "",
  append = FALSE
)
}
\arguments{
\item{x}{\code{\link{WordAssociation-class}}}

\item{from}{rank of the first (sorted) attracted forms printed}

\item{to}{rank of the last (sorted) attracted forms printed}

\item{threshold}{print only attracted forms with association strength above the threshold}

\item{types}{character. Print only the given types.}

\item{parts}{character. Print only the given subcorpora.}

\item{sort.by}{character. The name of one of the word association indicator available in the  \code{\link{WordAssociation-class}} object for sorting the lexical types in each subcorpora. By default, the first indicator (indicator.name(x)[1])}

\item{file}{character. If no empty, print the table in the given file}

\item{append}{logical. If the \code{file} argument is no empty, select erasing file or adding the table at the end of the file, preserving existing content.}
}
\description{
Pretty-print the word attractions measures in a \code{\link{WordAssociation-class}} object. The lexical types
are grouped by subcorpora and, in each subcorpora, print the types sorted according to the requested word association measure.
}
\examples{
data(robespierre)
res <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
print(res)
# or:
res
}
