% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{PseudoR2}
\alias{PseudoR2}
\title{PseudoR2}
\usage{
PseudoR2(x, which = NULL)
}
\arguments{
\item{x}{glm, polr, or multinom model object to be evaluated}

\item{which}{character, out of "McFadden","AldrichNelson", "McFaddenAdj", "Nagelkerke", "CoxSnell", "Effron", "McKelveyZavoina", "Tjur", and "all".}
}
\value{
value of desired statistic
}
\description{
This function allows users to access more R2 values (ex. McFadden's)
}
\examples{
usage
PseudoR2(x, which = NULL)

}
