% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.print.R
\name{show,WordAssociation-method}
\alias{show,WordAssociation-method}
\title{Show a \code{\link{WordAssociation-class}} object}
\usage{
\S4method{show}{WordAssociation}(object)
}
\arguments{
\item{object}{\code{\link{WordAssociation-class}}.}
}
\description{
Show a \code{\link{WordAssociation-class}} object
}
\examples{
data(robespierre)
res <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
show(res)
}
