% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.print.R
\name{summary,WordAssociation-method}
\alias{summary,WordAssociation-method}
\title{Print a summary of a \code{\link{WordAssociation-class}} object}
\usage{
\S4method{summary}{WordAssociation}(object)
}
\arguments{
\item{object}{\code{\link{WordAssociation-class}}.}
}
\description{
Print a summary of a \code{\link{WordAssociation-class}} object
}
\examples{
data(robespierre)
res <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
summary(res)
}
