% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{methods}
\name{types}
\alias{types}
\alias{types,WordAssociation-method}
\title{List of the types of a corpus}
\usage{
types(obj)

\S4method{types}{WordAssociation}(obj)
}
\arguments{
\item{obj}{a \code{\link{WordAssociation-class}} object.}
}
\value{
A character vector containing the list of the linguistic
types corresponding to the computed association measure.
}
\description{
List of the types of a corpus
}
\examples{
data(robespierre)
x <- wam(data=robespierre, measure = c("chisq", "loglikelihood"))
types(x)
}
