% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam-functions.R
\name{wam-function}
\alias{wam-function}
\alias{wam.jaccard}
\alias{wam.MI}
\alias{wam.frequency}
\alias{wam.loglikelihood}
\alias{wam.collostruction}
\alias{wam.specificities}
\alias{wam.z}
\alias{wam.t}
\alias{wam.chisq}
\alias{wam.ar}
\alias{wam.g}
\title{Calculate word association measure according to various fomula}
\usage{
wam.jaccard(N, n, K, k)

wam.MI(N, n, K, k)

wam.frequency(N, n, K, k)

wam.loglikelihood(N, n, K, k, p.value = FALSE, two.sided = TRUE)

wam.collostruction(N, n, K, k)

wam.specificities(N, n, K, k, method = "log")

wam.z(N, n, K, k, yates.correction = FALSE)

wam.t(N, n, K, k)

wam.chisq(
  N,
  n,
  K,
  k,
  yates.correction = TRUE,
  p.value = TRUE,
  two.sided = FALSE
)

wam.ar(N, n, K, k)

wam.g(N, n, K, k)
}
\arguments{
\item{N}{numeric vector, the total number of occurrences in the corpus}

\item{n}{numeric vector, The frequency of word 1}

\item{K}{numeric vector, The frequency of word 2}

\item{k}{numeric vector, The frequency of the cooccurrence or word 1 and word 2.}

\item{p.value}{length 1 numeric vector}

\item{two.sided}{length 1 numeric vector}

\item{method}{character: one of log, logscale, base, logscale, gap}

\item{yates.correction}{length 1 numeric vector}
}
\value{
a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength

a numeric vector of association strength
}
\description{
Word association measures give an information about the tendency of two
words to co-occurr with greater (or lesser) than chance frequency.

\verb{
 t-score =  prob(X=chemistry, Y=physics) - ( prob(X=chemistry) prob(Y=physics) )  
 ----------------------------------------------------------------------
   sqrt((1/T) prob(X=chemistry, Y=physics))
}
}
\examples{
data(ar)
attach(ar)
wam.ar(N, n, K, k)
}
\references{
Dunning, T. 1993.  « Accurate methods for the statistics of surprise and coincidence ».  In: \emph{Computational linguistics}.  19/1.  MIT Press, pp.  61-74

Stefanowitsch A. \& Gries St. Th. 2003 "Collostructions: Investigating the interaction of words and constructions", \emph{International Journal of Corpus Linguistics}, 8/2, 209-234. http://www.anglistik.uni-muenchen.de/personen/professoren/schmid/schmid_publ/collostructional-analysis.pdf

Lafon, P. (1980). « Sur la variabilité de la fréquence des formes dans un corpus ». In: \emph{Mots}. 1. pp. 127--165. \url{http://www.persee.fr/web/revues/home/prescript/article/mots_0243-6450_1980_num_1_1_1008}

Church, K., Gale, W., Hanks, P. & Hindle, D., 1991.  « Using Statistics in Lexical Analysis ».  In : Zernik, U. (ed.), \emph{Lexical Acquisition}.  Hillsdale, Lawrence Erlbaum Ass., pp. 115--164.

Church, K., Gale, W., Hanks, P. & Hindle, D., 1991.  «
Using Statistics in Lexical Analysis ».  In : Zernik, U. (ed.),
\emph{Lexical Acquisition}.  Hillsdale, Lawrence Erlbaum Ass., pp.

Kenneth Ward Church, Patrick Hanks (1990) "Word Association Norms,
Mutual Information, and Lexicography" \emph{Computational Linguistics},
16/1, pages 22-29.
\url{http://www.aclweb.org/anthology/P89-1010.pdf}
}
\seealso{
see \code{\link{wam}} for another, higher level interface.
}
\author{
Bernard Desgraupes

Sylvain Loiseau
}
