% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\name{wam}
\alias{wam}
\title{Compute association measure and construct a \code{\link{WordAssociation-class}} object.}
\usage{
wam(data, measure = "loglikelihood", ...)
}
\arguments{
\item{data}{: A data frame with the following columns:
N, n, K, k, types, parts: 
  \itemize{
    \item{N}{number of tokens in the corpus}
    
    \item{n}{number of tokens in the subcorpus}
    
    \item{K}{frequency of the type in the corpus}
    
    \item{k}{frequency of the type in the subcorpus}
    
    \item{types}{Optional : linguistic types}
    
    \item{parts}{Name of the subset of the corpus with which
      the attraction is measured}
    
  }}

\item{measure}{a function implementing statistical attraction measures. These functions must accept 'N', 'n', 'K' and 'k' arguments. See the help page \code{wam-functions}.}

\item{...}{argument provided to the function of the "measure" parameter.}
}
\value{
a \code{\link{WordAssociation-class}} object
}
\description{
Compute association measure and construct a \code{\link{WordAssociation-class}} object.
}
\examples{
data(robespierre)
wam(data=robespierre, measure = c("chisq", "loglikelihood"))
}
\seealso{
\code{\link{WordAssociation-class}}, \code{wam-functions}
}
