% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{happen}
\alias{happen}
\title{A contingency table lexem ~ construction}
\format{A 2*2 numeric matrix}
\usage{
data(happen)
}
\description{
A contingency table lexem ~ construction
  A 2*2 contingency table containing the frequency of lexem in the slot of a construction :
    \tabular{lrr}{
      \tab accident\tab !accident\cr
      N waiting to happen \tab       14\tab        21\cr
      !N waiting to happen\tab     8606\tab  10197659\cr
    }
}
\examples{
data(happen)
happen.vec <- cont2vec(happen)
happen.mar <- marginal(happen.vec)
attach(happen.mar)
wam.collostruction(N, n, K, k)
detach(happen.mar)
}
\references{
Stefanowitsch A. \& Gries St. Th. 2003 "Collostructions: Investigating the interaction of words and constructions", \emph{International Journal of Corpus Linguistics}, 8/2, 209-234, p. 219.
}
