\name{n-methods}
\docType{methods}
\alias{n}
\alias{n-methods}
\alias{n,WordAssociation-method}
\title{The size of the sub corpus in a word association list}
\description{

}
\section{Methods}{
\describe{
\item{\code{signature(obj = "WordAssociation")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{N}}, \code{\link{K}}, \code{\link{k}}, and \code{\link{WordAssociation}} for
  the big picture.
}

\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
