\name{wam}
\alias{wam}
\alias{wam-package}

\title{~ Overview: ... ~}
\description{
  \tabular{ll}{
    Version: \tab \Sexpr[stage=render]{library(utils); packageVersion("wam")}\cr
    Date: \tab \Sexpr[stage=build]{format(Sys.time(),"\%Y-\%m-\%d")}\cr
    License: \tab \Sexpr[stage=build]{library(utils); packageDescription("wam")$License} \cr
  }
}

\details{
  The package \code{wam} ...
}

\section{Author}{
  Bernard Desgraupes \cr
  \email{bernard.desgraupes@u-paris10.fr}\cr
  University of Paris Ouest - Nanterre - Lab Modal'X \cr \cr
}

\references{}

\seealso{}

\keyword{package}
