\name{WordAssociation}
\alias{WordAssociation}
\alias{WordAssociation-class}
\title{WordAssociation object}
\description{
See \code{\link{wam}} for constructing an WordAssociation object.

Contains word association measure indicators.
}
\details{
  For each element the following pieces of information are available :
 \itemize{
   \item K
   \item k
   \item N
   \item n
   \item type
   \item part
   \item one or several computed association measure indicators
   }

See the functions \code{\link{N}}, \code{\link{n}}, \code{\link{K}}, \code{\link{k}},
\code{\link{types}}, \code{\link{parts}}, \code{\link{indicator.name}} and \code{\link{association}}
for retrieving these pieces of information.
   
}
\author{
  Sylvain Loiseau
}
\seealso{
\code{\link{wam}}
}
