% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.object.R
\docType{class}
\name{WordAssociation-class}
\alias{WordAssociation-class}
\title{Contains word association measure indicators.}
\value{
a \code{\link{WordAssociation-class}} object
}
\description{
See \code{\link{wam}} for constructing a WordAssociation object.
}
\section{Slots}{

\describe{
\item{\code{N}}{numeric.}

\item{\code{n}}{numeric.}

\item{\code{K}}{numeric.}

\item{\code{k}}{numeric.}

\item{\code{association}}{matrix.}

\item{\code{indicator.name}}{character.}

\item{\code{types}}{character.}

\item{\code{parts}}{character. 

See the functions \code{\link{N-big}}, \code{\link{n-small}}, \code{\link{K-big}}, \code{\link{k-small}},
\code{\link{types}}, \code{\link{parts}}, \code{\link{indicator.name}} and \code{\link{association}}
for retrieving these pieces of information.}
}}

\seealso{
\code{\link{wam}}
}
\author{
Sylvain Loiseau
}
